Moved AQI code into separate file
